+++ kpops/const/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "10.4.2"
+__version__ = "10.5.0"
 KPOPS = "KPOps"
 KPOPS_MODULE = "kpops."